[refactor][jinj00oo]: 스크롤 트래킹 커스텀훅화 (#120)
