+++ src/app-pages/bylaws/ui/BylawsPage.tsx
@@ -5,6 +5,7 @@ import { bylawsData } from '@/app-pages/bylaws/model/data';
 import { useEffect, useRef } from 'react';
 import { trackBylawsEvent } from '@/features/bylaws/lib/trackBylawsEvent';
 import { BYLAWS_EVENTS } from '@/features/bylaws/model/types';
+import { useDynamicScrollTracking } from '@/shared/hooks/useDynamicScrollTracking';
 
 export default function BylawsPage() {
   // 페이지 진입 트래킹 (최초 1회)
@@ -16,63 +17,9 @@ export default function BylawsPage() {
   }, []);
 
   // 스크롤 퍼센트 트래킹
-  const scrollerRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const el = scrollerRef.current;
-    if (!el) return;
-
-    const THRESHOLDS = [0, 25, 50, 75, 100];
-    let sent = new Set<number>();
-    let prevScrollHeight = el.scrollHeight;
-
-    // 퍼센트 기준에 도달했을 때 이벤트 전송
-    const fire = (t: number) => {
-      if (!sent.has(t)) {
-        sent.add(t);
-        trackBylawsEvent(BYLAWS_EVENTS.SCROLL_RULES_PAGE, { percent: t });
-      }
-    };
-
-    // 현재 스크롤 퍼센트 계산
-    const getPercent = () => {
-      const max = el.scrollHeight - el.clientHeight;
-      if (max <= 0) return 0;
-
-      const ratio = el.scrollTop / max;
-      // 거의 끝까지 내렸다면 강제로 100으로 처리
-      if (ratio >= 0.99) return 100;
-
-      return Math.floor((el.scrollTop / max) * 100);
-    };
-
-    // 스크롤 시 퍼센트 체크
-    const onScroll = () => {
-      const p = getPercent();
-      THRESHOLDS.forEach((t) => {
-        if (p >= t && !sent.has(t)) fire(t);
-      });
-    };
-
-    // 초기 0% 트래킹
-    fire(0);
-    el.addEventListener('scroll', onScroll, { passive: true });
-
-    // scrollHeight 변화 감시 (아코디언 열림 등)
-    // Note: 페이지 높이 변경 시 sent를 재설정하여, 새로운 높이 기준으로 임계치를 다시 트래킹
-    const interval = setInterval(() => {
-      const cur = el.scrollHeight;
-      if (cur !== prevScrollHeight) {
-        prevScrollHeight = cur;
-        const p = getPercent();
-        sent = new Set(THRESHOLDS.filter((t) => t <= p));
-      }
-    }, 1000);
-
-    return () => {
-      el.removeEventListener('scroll', onScroll);
-      clearInterval(interval);
-    };
-  }, []);
+  const scrollerRef = useDynamicScrollTracking<HTMLDivElement>((percent) => {
+    trackBylawsEvent(BYLAWS_EVENTS.SCROLL_RULES_PAGE, { percent });
+  });
 
   return (
     <main
